commit package-lock.json after npm install on daine's laptop
the diffs seem to be sha1 (@5kb) vs sha512 (@dw2) integrity hashes
and a carrot (`\^`) before package versions from @5kb. I assume the
carrot means 'this version or higher'. I do not know why there is
a difference. We should investigate this to understand the reason
but for now it doesn't seem to be impacting the running of the app,
the tests, or the build.